+++ runtime/devnet/src/pallets/mod.rs
@@ -1,0 +1,2 @@
+pub mod system;
+pub use system::*;
